+++ modules/portal/app/views/v2client.scala.html
@@ -14,8 +14,8 @@
 		<link rel="stylesheet" type="text/css" id="theme" href="/public/gentelella/build/css/custom.css"/>
 		<link rel="stylesheet" type="text/css" id="theme" href="/public/css/vinyldns.css"/>
 		<link rel="stylesheet" type="text/css" id="theme" href="/public/css/theme-overrides.css"/>
-		<script src="@routes.Assets.versioned("v2client-fastopt-bundle.js")" type="text/javascript"></script>
-		<script src="@routes.Assets.versioned("v2client-opt-bundle.js")" type="text/javascript"></script>
+		<script src="@routes.Assets.versioned("client-fastopt-bundle.js")" type="text/javascript"></script>
+		<script src="@routes.Assets.versioned("client-opt-bundle.js")" type="text/javascript"></script>
 	</head>
 	<body onload="ReactApp.main('root')">
 		<div id="root"></div>
